Move all settings, rotation, and results functionality
