+++ psng/python/__init__.py
@@ -17,12 +17,16 @@
 
 from .jog import ProbeScreenJog
 from .psng import ProbeScreenClass
+from .settings import ProbeScreenSettings
 from .zero import ProbeScreenZero
+from .rotation import ProbeScreenRotation
 
 
 def get_handlers(halcomp, builder, useropts):
     return [
         ProbeScreenClass(halcomp, builder, useropts),
+        ProbeScreenSettings(halcomp, builder, useropts),
         ProbeScreenJog(halcomp, builder, useropts),
         ProbeScreenZero(halcomp, builder, useropts),
+        ProbeScreenRotation(halcomp, builder, useropts),
     ]
