+++ psng/python/psng.py
@@ -30,81 +30,6 @@ CONFIGPATH1 = os.environ["CONFIG_DIR"]
 
 
 class ProbeScreenClass(ProbeScreenBase):
-    # calculate corner coordinates in rotated coord. system
-    def calc_cross_rott(self, x1=0.0, y1=0.0, x2=0.0, y2=0.0, a1=0.0, a2=90.0):
-        coord = [0, 0]
-        k1 = math.tan(math.radians(a1))
-        k2 = math.tan(math.radians(a2))
-        coord[0] = (k1 * x1 - k2 * x2 + y2 - y1) / (k1 - k2)
-        coord[1] = k1 * (coord[0] - x1) + y1
-        return coord
-
-    # rotate point coordinates
-    def rott_point(self, x1=0.0, y1=0.0, a1=0.0):
-        coord = [x1, y1]
-        if a1 != 0:
-            if self.halcomp["set_zero"]:
-                xc = self.halcomp["ps_offs_x"]
-                yc = self.halcomp["ps_offs_y"]
-            else:
-                self.stat.poll()
-                xc = (
-                    self.stat.position[0]
-                    - self.stat.g5x_offset[0]
-                    - self.stat.g92_offset[0]
-                    - self.stat.tool_offset[0]
-                )
-                yc = (
-                    self.stat.position[1]
-                    - self.stat.g5x_offset[1]
-                    - self.stat.g92_offset[1]
-                    - self.stat.tool_offset[1]
-                )
-            t = math.radians(a1)
-            coord[0] = (x1 - xc) * math.cos(t) - (y1 - yc) * math.sin(t) + xc
-            coord[1] = (x1 - xc) * math.sin(t) + (y1 - yc) * math.cos(t) + yc
-        return coord
-
-    # rotate around 0,0 point coordinates
-    def rott00_point(self, x1=0.0, y1=0.0, a1=0.0):
-        coord = [x1, y1]
-        if a1 != 0:
-            t = math.radians(a1)
-            coord[0] = x1 * math.cos(t) - y1 * math.sin(t)
-            coord[1] = x1 * math.sin(t) + y1 * math.cos(t)
-        return coord
-
-    def probed_position_with_offsets(self):
-        self.stat.poll()
-        probed_position = list(self.stat.probed_position)
-        coord = list(self.stat.probed_position)
-        g5x_offset = list(self.stat.g5x_offset)
-        g92_offset = list(self.stat.g92_offset)
-        tool_offset = list(self.stat.tool_offset)
-        #        print "g5x_offset=",g5x_offset
-        #        print "g92_offset=",g92_offset
-        #        print "tool_offset=",tool_offset
-        #        print "actual position=",self.stat.actual_position
-        #        print "position=",self.stat.position
-        #        print "joint_actual position=",self.stat.joint_actual_position
-        #        print "joint_position=",self.stat.joint_position
-        #        print "probed position=",self.stat.probed_position
-        for i in range(0, len(probed_position) - 1):
-            coord[i] = (
-                probed_position[i] - g5x_offset[i] - g92_offset[i] - tool_offset[i]
-            )
-        angl = self.stat.rotation_xy
-        res = self.rott00_point(coord[0], coord[1], -angl)
-        coord[0] = res[0]
-        coord[1] = res[1]
-        return coord
-
-    # Auto Rot check
-    def on_chk_auto_rott_toggled(self, gtkcheckbutton, data=None):
-        self.halcomp["auto_rott"] = gtkcheckbutton.get_active()
-        self.hal_led_auto_rott.hal_pin.set(gtkcheckbutton.get_active())
-        self.prefs.putpref("chk_auto_rott", gtkcheckbutton.get_active(), bool)
-
     def set_zerro(self, s="XYZ", x=0.0, y=0.0, z=0.0):
         if self.halcomp["set_zero"]:
             #  Z current position
@@ -129,201 +54,33 @@ class ProbeScreenClass(ProbeScreenBase):
             self.gcode(c)
             time.sleep(1)
 
-    def rotate_coord_system(self, a=0.0):
-        self.spbtn_offs_angle.set_value(a)
-        self.lb_probe_a.set_text("%.3f" % a)
-        if self.chk_auto_rott.get_active():
-            s = "G10 L2 P0"
-            if self.halcomp["set_zero"]:
-                s += " X%s" % self.halcomp["ps_offs_x"]
-                s += " Y%s" % self.halcomp["ps_offs_y"]
-            else:
-                self.stat.poll()
-                x = self.stat.position[0]
-                y = self.stat.position[1]
-                s += " X%s" % x
-                s += " Y%s" % y
-            s += " R%s" % a
-            self.gcode(s)
-            time.sleep(1)
-
-    # Spin  buttons
-
-    def on_spbtn1_search_vel_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        #        print "Key %s (%d) was pressed" % (keyname, data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn1_probe_vel_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn1_probe_max_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn1_probe_latch_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn1_probe_diam_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn1_xy_clearance_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn1_edge_lenght_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn1_z_clearance_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn_offs_angle_key_press_event(self, gtkspinbutton, data=None):
-        keyname = gtk.gdk.keyval_name(data.keyval)
-        if keyname == "Return":
-            gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        else:
-            gtkspinbutton.modify_font(pango.FontDescription("italic"))
-
-    def on_spbtn1_search_vel_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal "))
-        self.halcomp["ps_searchvel"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_searchvel", gtkspinbutton.get_value(), float)
-
-    def on_spbtn1_probe_vel_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        self.halcomp["ps_probevel"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_probevel", gtkspinbutton.get_value(), float)
-
-    def on_spbtn1_probe_max_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        self.halcomp["ps_probe_max"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_probe_max", gtkspinbutton.get_value(), float)
-
-    def on_spbtn1_probe_latch_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        self.halcomp["ps_probe_latch"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_probe_latch", gtkspinbutton.get_value(), float)
-
-    def on_spbtn1_probe_diam_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        self.halcomp["ps_probe_diam"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_probe_diam", gtkspinbutton.get_value(), float)
-
-    def on_spbtn1_xy_clearance_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        self.halcomp["ps_xy_clearance"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_xy_clearance", gtkspinbutton.get_value(), float)
-
-    def on_spbtn1_edge_lenght_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        self.halcomp["ps_edge_lenght"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_edge_lenght", gtkspinbutton.get_value(), float)
-
-    def on_spbtn1_z_clearance_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        self.halcomp["ps_z_clearance"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_z_clearance", gtkspinbutton.get_value(), float)
-
-    def on_spbtn_offs_angle_value_changed(self, gtkspinbutton, data=None):
-        gtkspinbutton.modify_font(pango.FontDescription("normal"))
-        self.halcomp["ps_offs_angle"] = gtkspinbutton.get_value()
-        self.prefs.putpref("ps_offs_angle", gtkspinbutton.get_value(), float)
-
-    def z_clearance_down(self, data=None):
-        # move Z - z_clearance
-        s = """G91
-        G1 Z-%f
-        G90""" % (
-            self.spbtn1_z_clearance.get_value()
-        )
-        if self.gcode(s) == -1:
-            return -1
-        return 0
-
-    def z_clearance_up(self, data=None):
-        # move Z + z_clearance
-        s = """G91
-        G1 Z%f
-        G90""" % (
-            self.spbtn1_z_clearance.get_value()
-        )
-        if self.gcode(s) == -1:
-            return -1
-        return 0
-
-    def lenght_x(self, data=None):
+    def lenght_x(self):
+        # TODO: Rework this to no need access to self._lb_probe_*
         res = 0
-        if self.lb_probe_xm.get_text() == "" or self.lb_probe_xp.get_text() == "":
+        if self._lb_probe_xm.get_text() == "" or self._lb_probe_xp.get_text() == "":
             return res
-        xm = float(self.lb_probe_xm.get_text())
-        xp = float(self.lb_probe_xp.get_text())
+        xm = float(self._lb_probe_xm.get_text())
+        xp = float(self._lb_probe_xp.get_text())
         if xm < xp:
             res = xp - xm
         else:
             res = xm - xp
-        self.lb_probe_lx.set_text("%.4f" % res)
+        self.display_result_lx(res)
         return res
 
-    def lenght_y(self, data=None):
+    def lenght_y(self):
+        # TODO: Rework this to no need access to self._lb_probe_*
         res = 0
-        if self.lb_probe_ym.get_text() == "" or self.lb_probe_yp.get_text() == "":
+        if self._lb_probe_ym.get_text() == "" or self._lb_probe_yp.get_text() == "":
             return res
-        ym = float(self.lb_probe_ym.get_text())
-        yp = float(self.lb_probe_yp.get_text())
+        ym = float(self._lb_probe_ym.get_text())
+        yp = float(self._lb_probe_yp.get_text())
         if ym < yp:
             res = yp - ym
         else:
             res = ym - yp
-        self.lb_probe_ly.set_text("%.4f" % res)
+        self.display_result_ly(res)
         return res
-
-    # --------------  Rotate buttons -----------------
-    def on_btn_set_angle_released(self, gtkbutton, data=None):
-        self.prefs.putpref("ps_offs_angle", self.spbtn_offs_angle.get_value(), float)
-        self.lb_probe_a.set_text("%.3f" % self.spbtn_offs_angle.get_value())
-        s = "G10 L2 P0"
-        if self.halcomp["set_zero"]:
-            s += " X%.4f" % self.halcomp["ps_offs_x"]
-            s += " Y%.4f" % self.halcomp["ps_offs_y"]
-        else:
-            self.stat.poll()
-            x = self.stat.position[0]
-            y = self.stat.position[1]
-            s += " X%.4f" % x
-            s += " Y%.4f" % y
-        s += " R%.4f" % self.spbtn_offs_angle.get_value()
-        print("s=", s)
-        self.gcode(s)
-        time.sleep(1)
 
     # --------------  Command buttons -----------------
     #               Measurement outside
@@ -334,7 +91,7 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X-%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value()
+            self.halcomp["ps_xy_clearance"]
         )
         if self.gcode(s) == -1:
             return
@@ -344,8 +101,8 @@ class ProbeScreenClass(ProbeScreenBase):
         if self.ocode("o<psng_xplus> call") == -1:
             return
         a = self.probed_position_with_offsets()
-        xres = float(a[0] + 0.5 * self.spbtn1_probe_diam.get_value())
-        self.lb_probe_xp.set_text("%.4f" % xres)
+        xres = float(a[0] + 0.5 * self.halcomp["ps_probe_diam"])
+        self.display_result_xp(xres)
         self.lenght_x()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -377,7 +134,7 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 Y-%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value()
+            self.halcomp["ps_xy_clearance"]
         )
         if self.gcode(s) == -1:
             return
@@ -387,8 +144,8 @@ class ProbeScreenClass(ProbeScreenBase):
         if self.ocode("o<psng_yplus> call") == -1:
             return
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % yres)
+        yres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -420,7 +177,7 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value()
+            self.halcomp["ps_xy_clearance"]
         )
         if self.gcode(s) == -1:
             return
@@ -430,8 +187,8 @@ class ProbeScreenClass(ProbeScreenBase):
         if self.ocode("o<psng_xminus> call") == -1:
             return
         a = self.probed_position_with_offsets()
-        xres = float(a[0] - 0.5 * self.spbtn1_probe_diam.get_value())
-        self.lb_probe_xm.set_text("%.4f" % xres)
+        xres = float(a[0] - 0.5 * self.halcomp["ps_probe_diam"])
+        self.display_result_xm(xres)
         self.lenght_x()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -463,7 +220,7 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 Y%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value()
+            self.halcomp["ps_xy_clearance"]
         )
         if self.gcode(s) == -1:
             return
@@ -473,8 +230,8 @@ class ProbeScreenClass(ProbeScreenBase):
         if self.ocode("o<psng_yminus> call") == -1:
             return
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % yres)
+        yres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -508,8 +265,8 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X-%f Y%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value(),
-            self.spbtn1_edge_lenght.get_value(),
+            self.halcomp["ps_xy_clearance"],
+            self.halcomp["ps_edge_lenght"],
         )
         if self.gcode(s) == -1:
             return
@@ -520,17 +277,15 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xres = float(a[0] + 0.5 * self.spbtn1_probe_diam.get_value())
-        self.lb_probe_xp.set_text("%.4f" % xres)
+        xres = float(a[0] + 0.5 * self.halcomp["ps_probe_diam"])
+        self.display_result_xp(xres)
         self.lenght_x()
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
 
         # move X + edge_lenght +xy_clearance,  Y - edge_lenght - xy_clearance
-        tmpxy = (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpxy = self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X%f Y-%f
         G90""" % (
@@ -546,8 +301,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % yres)
+        yres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -579,8 +334,8 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X-%f Y-%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value(),
-            self.spbtn1_edge_lenght.get_value(),
+            self.halcomp["ps_xy_clearance"],
+            self.halcomp["ps_edge_lenght"],
         )
         if self.gcode(s) == -1:
             return
@@ -591,17 +346,15 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xres = float(a[0] + 0.5 * self.spbtn1_probe_diam.get_value())
-        self.lb_probe_xp.set_text("%.4f" % xres)
+        xres = float(a[0] + 0.5 * self.halcomp["ps_probe_diam"])
+        self.display_result_xp(xres)
         self.lenght_x()
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
 
         # move X + edge_lenght +xy_clearance,  Y + edge_lenght + xy_clearance
-        tmpxy = (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpxy = self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X%f Y%f
         G90""" % (
@@ -617,8 +370,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % yres)
+        yres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(yres)
         self.add_history(
             gtkbutton.get_tooltip_text(),
             "XpLxYmLy",
@@ -649,8 +402,8 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X%f Y%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value(),
-            self.spbtn1_edge_lenght.get_value(),
+            self.halcomp["ps_xy_clearance"],
+            self.halcomp["ps_edge_lenght"],
         )
         if self.gcode(s) == -1:
             return
@@ -661,17 +414,15 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xres = float(a[0] - 0.5 * self.spbtn1_probe_diam.get_value())
-        self.lb_probe_xm.set_text("%.4f" % xres)
+        xres = float(a[0] - 0.5 * self.halcomp["ps_probe_diam"])
+        self.display_result_xm(xres)
         self.lenght_x()
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
 
         # move X - edge_lenght - xy_clearance,  Y - edge_lenght - xy_clearance
-        tmpxy = (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpxy = self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X-%f Y-%f
         G90""" % (
@@ -687,8 +438,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % yres)
+        yres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -720,8 +471,8 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X%f Y-%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value(),
-            self.spbtn1_edge_lenght.get_value(),
+            self.halcomp["ps_xy_clearance"],
+            self.halcomp["ps_edge_lenght"],
         )
         if self.gcode(s) == -1:
             return
@@ -732,17 +483,15 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xres = float(a[0] - 0.5 * self.spbtn1_probe_diam.get_value())
-        self.lb_probe_xm.set_text("%.4f" % xres)
+        xres = float(a[0] - 0.5 * self.halcomp["ps_probe_diam"])
+        self.display_result_xm(xres)
         self.lenght_x()
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
 
         # move X - edge_lenght - xy_clearance,  Y + edge_lenght + xy_clearance
-        tmpxy = (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpxy = self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X-%f Y%f
         G90""" % (
@@ -758,8 +507,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % yres)
+        yres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -788,9 +537,7 @@ class ProbeScreenClass(ProbeScreenBase):
     # Center X+ X- Y+ Y-
     def on_xy_center_released(self, gtkbutton, data=None):
         # move X - edge_lenght- xy_clearance
-        tmpx = (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpx = self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X-%f
         G90""" % (
@@ -805,16 +552,14 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xpres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xp.set_text("%.4f" % xpres)
+        xpres = float(a[0]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xp(xpres)
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
 
         # move X + 2 edge_lenght + 2 xy_clearance
-        tmpx = 2 * (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpx = 2 * (self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"])
         s = """G91
         G1 X%f
         G90""" % (
@@ -830,11 +575,11 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xmres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xm.set_text("%.4f" % xmres)
+        xmres = float(a[0]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xm(xmres)
         self.lenght_x()
         xcres = 0.5 * (xpres + xmres)
-        self.lb_probe_xc.set_text("%.4f" % xcres)
+        self.display_result_xc(xcres)
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
@@ -846,9 +591,7 @@ class ProbeScreenClass(ProbeScreenBase):
             return
 
         # move Y - edge_lenght- xy_clearance
-        tmpy = (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpy = self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 Y-%f
         G90""" % (
@@ -863,16 +606,14 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ypres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % ypres)
+        ypres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(ypres)
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
 
         # move Y + 2 edge_lenght + 2 xy_clearance
-        tmpy = 2 * (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpy = 2 * (self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"])
         s = """G91
         G1 Y%f
         G90""" % (
@@ -887,14 +628,14 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ymres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % ymres)
+        ymres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(ymres)
         self.lenght_y()
         # find, show and move to finded  point
         ycres = 0.5 * (ypres + ymres)
-        self.lb_probe_yc.set_text("%.4f" % ycres)
+        self.display_result_yc(ycres)
         diam = ymres - ypres
-        self.lb_probe_d.set_text("%.4f" % diam)
+        self.display_result_d(diam)
         self.add_history(
             gtkbutton.get_tooltip_text(),
             "XmXcXpLxYmYcYpLyD",
@@ -931,8 +672,8 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X-%f Y-%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value(),
-            self.spbtn1_edge_lenght.get_value(),
+            self.halcomp["ps_xy_clearance"],
+            self.halcomp["ps_edge_lenght"],
         )
         if self.gcode(s) == -1:
             return
@@ -943,14 +684,12 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xp.set_text("%.4f" % xres)
+        xres = float(a[0]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xp(xres)
         self.lenght_x()
 
         # move X - edge_lenght Y - xy_clearance
-        tmpxy = (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpxy = self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X-%f Y%f
         G90""" % (
@@ -964,8 +703,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % yres)
+        yres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -997,8 +736,8 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X-%f Y%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value(),
-            self.spbtn1_edge_lenght.get_value(),
+            self.halcomp["ps_xy_clearance"],
+            self.halcomp["ps_edge_lenght"],
         )
         if self.gcode(s) == -1:
             return
@@ -1009,14 +748,12 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xp.set_text("%.4f" % xres)
+        xres = float(a[0]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xp(xres)
         self.lenght_x()
 
         # move X - edge_lenght Y + xy_clearance
-        tmpxy = (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpxy = self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X-%f Y-%f
         G90""" % (
@@ -1030,8 +767,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % yres)
+        yres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -1063,8 +800,8 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X%f Y-%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value(),
-            self.spbtn1_edge_lenght.get_value(),
+            self.halcomp["ps_xy_clearance"],
+            self.halcomp["ps_edge_lenght"],
         )
         if self.gcode(s) == -1:
             return
@@ -1075,14 +812,12 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xm.set_text("%.4f" % xres)
+        xres = float(a[0]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xm(xres)
         self.lenght_x()
 
         # move X + edge_lenght Y - xy_clearance
-        tmpxy = (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpxy = self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X%f Y%f
         G90""" % (
@@ -1097,8 +832,8 @@ class ProbeScreenClass(ProbeScreenBase):
 
         # show Y result
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % yres)
+        yres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -1130,8 +865,8 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X%f Y%f
         G90""" % (
-            self.spbtn1_xy_clearance.get_value(),
-            self.spbtn1_edge_lenght.get_value(),
+            self.halcomp["ps_xy_clearance"],
+            self.halcomp["ps_edge_lenght"],
         )
         if self.gcode(s) == -1:
             return
@@ -1142,14 +877,12 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xm.set_text("%.4f" % xres)
+        xres = float(a[0]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xm(xres)
         self.lenght_x()
 
         # move X + edge_lenght Y - xy_clearance
-        tmpxy = (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpxy = self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X%f Y-%f
         G90""" % (
@@ -1163,8 +896,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        yres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % yres)
+        yres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(yres)
         self.lenght_y()
         self.add_history(
             gtkbutton.get_tooltip_text(),
@@ -1195,9 +928,7 @@ class ProbeScreenClass(ProbeScreenBase):
         if self.z_clearance_down() == -1:
             return
         # move X - edge_lenght Y + xy_clearance
-        tmpx = (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpx = self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X-%f
         G90""" % (
@@ -1210,13 +941,11 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xmres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xm.set_text("%.4f" % xmres)
+        xmres = float(a[0]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xm(xmres)
 
         # move X +2 edge_lenght - 2 xy_clearance
-        tmpx = 2 * (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpx = 2 * (self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"])
         s = """G91
         G1 X%f
         G90""" % (
@@ -1229,11 +958,11 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xpres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xp.set_text("%.4f" % xpres)
+        xpres = float(a[0]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xp(xpres)
         self.lenght_x()
         xcres = 0.5 * (xmres + xpres)
-        self.lb_probe_xc.set_text("%.4f" % xcres)
+        self.display_result_xc(xcres)
 
         # move X to new center
         s = """G1 X%f""" % (xcres)
@@ -1241,9 +970,7 @@ class ProbeScreenClass(ProbeScreenBase):
             return
 
         # move Y - edge_lenght + xy_clearance
-        tmpy = (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpy = self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 Y-%f
         G90""" % (
@@ -1256,13 +983,11 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ymres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % ymres)
+        ymres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(ymres)
 
         # move Y +2 edge_lenght - 2 xy_clearance
-        tmpy = 2 * (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpy = 2 * (self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"])
         s = """G91
         G1 Y%f
         G90""" % (
@@ -1275,14 +1000,14 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ypres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % ypres)
+        ypres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(ypres)
         self.lenght_y()
         # find, show and move to finded  point
         ycres = 0.5 * (ymres + ypres)
-        self.lb_probe_yc.set_text("%.4f" % ycres)
+        self.display_result_yc(ycres)
         diam = 0.5 * ((xpres - xmres) + (ypres - ymres))
-        self.lb_probe_d.set_text("%.4f" % diam)
+        self.display_result_d(diam)
         self.add_history(
             gtkbutton.get_tooltip_text(),
             "XmXcXpLxYmYcYpLyD",
@@ -1308,305 +1033,6 @@ class ProbeScreenClass(ProbeScreenBase):
         self.set_zerro("XY")
 
     # --------------  Command buttons -----------------
-    #               Measurement angle
-    # -------------------------------------------------
-
-    # Angle
-    # Move Probe manual under corner 2-3 mm
-    # Y+Y+
-    def on_angle_yp_released(self, gtkbutton, data=None):
-        self.stat.poll()
-        xstart = (
-            self.stat.position[0]
-            - self.stat.g5x_offset[0]
-            - self.stat.g92_offset[0]
-            - self.stat.tool_offset[0]
-        )
-        # move Y - xy_clearance
-        s = """G91
-        G1 Y-%f
-        G90""" % (
-            self.spbtn1_xy_clearance.get_value()
-        )
-        if self.gcode(s) == -1:
-            return
-        if self.z_clearance_down() == -1:
-            return
-        # Start psng_yplus.ngc
-        if self.ocode("o<psng_yplus> call") == -1:
-            return
-        # show Y result
-        a = self.probed_position_with_offsets()
-        ycres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yc.set_text("%.4f" % ycres)
-        # move Z to start point
-        if self.z_clearance_up() == -1:
-            return
-        # move X + edge_lenght
-        s = """G91
-        G1 X%f
-        G90""" % (
-            self.spbtn1_edge_lenght.get_value()
-        )
-        if self.gcode(s) == -1:
-            return
-        if self.z_clearance_down() == -1:
-            return
-        # Start psng_yplus.ngc
-        if self.ocode("o<psng_yplus> call") == -1:
-            return
-        # show Y result
-        a = self.probed_position_with_offsets()
-        ypres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % ypres)
-        alfa = math.degrees(
-            math.atan2(ypres - ycres, self.spbtn1_edge_lenght.get_value())
-        )
-        self.add_history(
-            gtkbutton.get_tooltip_text(),
-            "YcYpA",
-            0,
-            0,
-            0,
-            0,
-            0,
-            ycres,
-            ypres,
-            0,
-            0,
-            0,
-            alfa,
-        )
-
-        # move Z to start point
-        if self.z_clearance_up() == -1:
-            return
-        # move XY to adj start point
-        s = "G1 X%f Y%f" % (xstart, ycres)
-        if self.gcode(s) == -1:
-            return
-        self.rotate_coord_system(alfa)
-
-    # Y-Y-
-    def on_angle_ym_released(self, gtkbutton, data=None):
-        self.stat.poll()
-        xstart = (
-            self.stat.position[0]
-            - self.stat.g5x_offset[0]
-            - self.stat.g92_offset[0]
-            - self.stat.tool_offset[0]
-        )
-        # move Y + xy_clearance
-        s = """G91
-        G1 Y%f
-        G90""" % (
-            self.spbtn1_xy_clearance.get_value()
-        )
-        if self.gcode(s) == -1:
-            return
-        if self.z_clearance_down() == -1:
-            return
-        # Start psng_yminus.ngc
-        if self.ocode("o<psng_yminus> call") == -1:
-            return
-        # show Y result
-        a = self.probed_position_with_offsets()
-        ycres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yc.set_text("%.4f" % ycres)
-        # move Z to start point
-        if self.z_clearance_up() == -1:
-            return
-        # move X - edge_lenght
-        s = """G91
-        G1 X-%f
-        G90""" % (
-            self.spbtn1_edge_lenght.get_value()
-        )
-        if self.gcode(s) == -1:
-            return
-        if self.z_clearance_down() == -1:
-            return
-        # Start psng_yminus.ngc
-        if self.ocode("o<psng_yminus> call") == -1:
-            return
-        # show Y result
-        a = self.probed_position_with_offsets()
-        ymres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % ymres)
-        alfa = math.degrees(
-            math.atan2(ycres - ymres, self.spbtn1_edge_lenght.get_value())
-        )
-        self.add_history(
-            gtkbutton.get_tooltip_text(),
-            "YmYcA",
-            0,
-            0,
-            0,
-            0,
-            ymres,
-            ycres,
-            0,
-            0,
-            0,
-            0,
-            alfa,
-        )
-        # move Z to start point
-        if self.z_clearance_up() == -1:
-            return
-        # move XY to adj start point
-        s = "G1 X%f Y%f" % (xstart, ycres)
-        if self.gcode(s) == -1:
-            return
-        self.rotate_coord_system(alfa)
-
-    # X+X+
-    def on_angle_xp_released(self, gtkbutton, data=None):
-        self.stat.poll()
-        ystart = (
-            self.stat.position[1]
-            - self.stat.g5x_offset[1]
-            - self.stat.g92_offset[1]
-            - self.stat.tool_offset[1]
-        )
-        # move X - xy_clearance
-        s = """G91
-        G1 X-%f
-        G90""" % (
-            self.spbtn1_xy_clearance.get_value()
-        )
-        if self.gcode(s) == -1:
-            return
-        if self.z_clearance_down() == -1:
-            return
-        # Start psng_xplus.ngc
-        if self.ocode("o<psng_xplus> call") == -1:
-            return
-        # show X result
-        a = self.probed_position_with_offsets()
-        xcres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xc.set_text("%.4f" % xcres)
-        # move Z to start point
-        if self.z_clearance_up() == -1:
-            return
-        # move Y - edge_lenght
-        s = """G91
-        G1 Y-%f
-        G90""" % (
-            self.spbtn1_edge_lenght.get_value()
-        )
-        if self.gcode(s) == -1:
-            return
-        if self.z_clearance_down() == -1:
-            return
-        # Start psng_xplus.ngc
-        if self.ocode("o<psng_xplus> call") == -1:
-            return
-        # show X result
-        a = self.probed_position_with_offsets()
-        xpres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xp.set_text("%.4f" % xpres)
-        alfa = math.degrees(
-            math.atan2(xcres - xpres, self.spbtn1_edge_lenght.get_value())
-        )
-        self.add_history(
-            gtkbutton.get_tooltip_text(),
-            "XcXpA",
-            0,
-            xcres,
-            xpres,
-            0,
-            0,
-            0,
-            0,
-            0,
-            0,
-            0,
-            alfa,
-        )
-        # move Z to start point
-        if self.z_clearance_up() == -1:
-            return
-        # move XY to adj start point
-        s = "G1 X%f Y%f" % (xcres, ystart)
-        if self.gcode(s) == -1:
-            return
-        self.rotate_coord_system(alfa)
-
-    # X-X-
-    def on_angle_xm_released(self, gtkbutton, data=None):
-        self.stat.poll()
-        ystart = (
-            self.stat.position[1]
-            - self.stat.g5x_offset[1]
-            - self.stat.g92_offset[1]
-            - self.stat.tool_offset[1]
-        )
-        # move X + xy_clearance
-        s = """G91
-        G1 X%f
-        G90""" % (
-            self.spbtn1_xy_clearance.get_value()
-        )
-        if self.gcode(s) == -1:
-            return
-        if self.z_clearance_down() == -1:
-            return
-        # Start psng_xminus.ngc
-        if self.ocode("o<psng_xminus> call") == -1:
-            return
-        # show X result
-        a = self.probed_position_with_offsets()
-        xcres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xc.set_text("%.4f" % xcres)
-        # move Z to start point
-        if self.z_clearance_up() == -1:
-            return
-        # move Y + edge_lenght
-        s = """G91
-        G1 Y%f
-        G90""" % (
-            self.spbtn1_edge_lenght.get_value()
-        )
-        if self.gcode(s) == -1:
-            return
-        if self.z_clearance_down() == -1:
-            return
-        # Start psng_xminus.ngc
-        if self.ocode("o<psng_xminus> call") == -1:
-            return
-        # show X result
-        a = self.probed_position_with_offsets()
-        xmres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xm.set_text("%.4f" % xmres)
-        alfa = math.degrees(
-            math.atan2(xcres - xmres, self.spbtn1_edge_lenght.get_value())
-        )
-        self.add_history(
-            gtkbutton.get_tooltip_text(),
-            "XmXcA",
-            xmres,
-            xcres,
-            0,
-            0,
-            0,
-            0,
-            0,
-            0,
-            0,
-            0,
-            alfa,
-        )
-        # move Z to start point
-        if self.z_clearance_up() == -1:
-            return
-        # move XY to adj start point
-        s = "G1 X%f Y%f" % (xcres, ystart)
-        if self.gcode(s) == -1:
-            return
-        self.rotate_coord_system(alfa)
-
-    # --------------  Command buttons -----------------
     #               Measurement LENGHT
     # -------------------------------------------------
     # LENGHT
@@ -1614,9 +1040,7 @@ class ProbeScreenClass(ProbeScreenBase):
     # Lx OUT
     def on_lx_out_released(self, gtkbutton, data=None):
         # move X - edge_lenght- xy_clearance
-        tmpx = (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpx = self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X-%f
         G90""" % (
@@ -1631,8 +1055,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xpres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xp.set_text("%.4f" % xpres)
+        xpres = float(a[0]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xp(xpres)
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
@@ -1642,10 +1066,7 @@ class ProbeScreenClass(ProbeScreenBase):
             return
 
         # move X + 2 edge_lenght +  xy_clearance
-        tmpx = (
-            2 * self.spbtn1_edge_lenght.get_value()
-            + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpx = 2 * self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X%f
         G90""" % (
@@ -1660,11 +1081,11 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xmres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xm.set_text("%.4f" % xmres)
+        xmres = float(a[0]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xm(xmres)
         self.lenght_x()
         xcres = 0.5 * (xpres + xmres)
-        self.lb_probe_xc.set_text("%.4f" % xcres)
+        self.display_result_xc(xcres)
         self.add_history(
             gtkbutton.get_tooltip_text(),
             "XmXcXpLx",
@@ -1692,9 +1113,7 @@ class ProbeScreenClass(ProbeScreenBase):
     # Ly OUT
     def on_ly_out_released(self, gtkbutton, data=None):
         # move Y - edge_lenght- xy_clearance
-        tmpy = (
-            self.spbtn1_edge_lenght.get_value() + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpy = self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 Y-%f
         G90""" % (
@@ -1709,8 +1128,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ypres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % ypres)
+        ypres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(ypres)
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
@@ -1720,10 +1139,7 @@ class ProbeScreenClass(ProbeScreenBase):
             return
 
         # move Y + 2 edge_lenght +  xy_clearance
-        tmpy = (
-            2 * self.spbtn1_edge_lenght.get_value()
-            + self.spbtn1_xy_clearance.get_value()
-        )
+        tmpy = 2 * self.halcomp["ps_edge_lenght"] + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 Y%f
         G90""" % (
@@ -1738,12 +1154,12 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ymres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % ymres)
+        ymres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(ymres)
         self.lenght_y()
         # find, show and move to finded  point
         ycres = 0.5 * (ypres + ymres)
-        self.lb_probe_yc.set_text("%.4f" % ycres)
+        self.display_result_yc(ycres)
         self.add_history(
             gtkbutton.get_tooltip_text(),
             "YmYcYpLy",
@@ -1773,9 +1189,7 @@ class ProbeScreenClass(ProbeScreenBase):
         if self.z_clearance_down() == -1:
             return
         # move X - edge_lenght Y + xy_clearance
-        tmpx = (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpx = self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X-%f
         G90""" % (
@@ -1788,13 +1202,11 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xmres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xm.set_text("%.4f" % xmres)
+        xmres = float(a[0]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xm(xmres)
 
         # move X +2 edge_lenght - 2 xy_clearance
-        tmpx = 2 * (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpx = 2 * (self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"])
         s = """G91
         G1 X%f
         G90""" % (
@@ -1807,11 +1219,11 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xpres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_xp.set_text("%.4f" % xpres)
+        xpres = float(a[0]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_xp(xpres)
         self.lenght_x()
         xcres = 0.5 * (xmres + xpres)
-        self.lb_probe_xc.set_text("%.4f" % xcres)
+        self.display_result_xc(xcres)
         self.add_history(
             gtkbutton.get_tooltip_text(),
             "XmXcXpLx",
@@ -1841,9 +1253,7 @@ class ProbeScreenClass(ProbeScreenBase):
         if self.z_clearance_down() == -1:
             return
         # move Y - edge_lenght + xy_clearance
-        tmpy = (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpy = self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 Y-%f
         G90""" % (
@@ -1856,13 +1266,11 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ymres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_ym.set_text("%.4f" % ymres)
+        ymres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_ym(ymres)
 
         # move Y +2 edge_lenght - 2 xy_clearance
-        tmpy = 2 * (
-            self.spbtn1_edge_lenght.get_value() - self.spbtn1_xy_clearance.get_value()
-        )
+        tmpy = 2 * (self.halcomp["ps_edge_lenght"] - self.halcomp["ps_xy_clearance"])
         s = """G91
         G1 Y%f
         G90""" % (
@@ -1875,12 +1283,12 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ypres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        self.lb_probe_yp.set_text("%.4f" % ypres)
+        ypres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
+        self.display_result_yp(ypres)
         self.lenght_y()
         # find, show and move to finded  point
         ycres = 0.5 * (ymres + ypres)
-        self.lb_probe_yc.set_text("%.4f" % ycres)
+        self.display_result_yc(ycres)
         self.add_history(
             gtkbutton.get_tooltip_text(),
             "YmYcYpLy",
@@ -1915,7 +1323,7 @@ class ProbeScreenClass(ProbeScreenBase):
         s = """G91
         G1 X-%f
         G90""" % (
-            0.5 * self.tsdiam + self.spbtn1_xy_clearance.get_value()
+            0.5 * self.tsdiam + self.halcomp["ps_xy_clearance"]
         )
         if self.gcode(s) == -1:
             return
@@ -1926,8 +1334,8 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xpres = float(a[0]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        #        self.lb_probe_xp.set_text( "%.4f" % xpres )
+        xpres = float(a[0]) + 0.5 * self.halcomp["ps_probe_diam"]
+
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
@@ -1937,7 +1345,7 @@ class ProbeScreenClass(ProbeScreenBase):
             return
 
         # move X + tsdiam +  xy_clearance
-        aa = self.tsdiam + self.spbtn1_xy_clearance.get_value()
+        aa = self.tsdiam + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 X%f
         G90""" % (
@@ -1953,11 +1361,10 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show X result
         a = self.probed_position_with_offsets()
-        xmres = float(a[0]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        #        self.lb_probe_xm.set_text( "%.4f" % xmres )
+        xmres = float(a[0]) - 0.5 * self.halcomp["ps_probe_diam"]
         self.lenght_x()
         xcres = 0.5 * (xpres + xmres)
-        self.lb_probe_xc.set_text("%.4f" % xcres)
+        self.display_result_xc(xcres)
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
@@ -1967,7 +1374,7 @@ class ProbeScreenClass(ProbeScreenBase):
             return
 
         # move Y - tsdiam/2 - xy_clearance
-        a = 0.5 * self.tsdiam + self.spbtn1_xy_clearance.get_value()
+        a = 0.5 * self.tsdiam + self.halcomp["ps_xy_clearance"]
         s = (
             """G91
         G1 Y-%f
@@ -1983,8 +1390,7 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ypres = float(a[1]) + 0.5 * self.spbtn1_probe_diam.get_value()
-        #        self.lb_probe_yp.set_text( "%.4f" % ypres )
+        ypres = float(a[1]) + 0.5 * self.halcomp["ps_probe_diam"]
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
@@ -1994,7 +1400,7 @@ class ProbeScreenClass(ProbeScreenBase):
             return
 
         # move Y + tsdiam +  xy_clearance
-        aa = self.tsdiam + self.spbtn1_xy_clearance.get_value()
+        aa = self.tsdiam + self.halcomp["ps_xy_clearance"]
         s = """G91
         G1 Y%f
         G90""" % (
@@ -2009,15 +1415,14 @@ class ProbeScreenClass(ProbeScreenBase):
             return
         # show Y result
         a = self.probed_position_with_offsets()
-        ymres = float(a[1]) - 0.5 * self.spbtn1_probe_diam.get_value()
-        #        self.lb_probe_ym.set_text( "%.4f" % ymres )
+        ymres = float(a[1]) - 0.5 * self.halcomp["ps_probe_diam"]
         self.lenght_y()
         # find, show and move to finded  point
         ycres = 0.5 * (ypres + ymres)
-        self.lb_probe_yc.set_text("%.4f" % ycres)
-        diam = self.spbtn1_probe_diam.get_value() + (ymres - ypres - self.tsdiam)
+        self.display_result_yc(ycres)
+        diam = self.halcomp["ps_probe_diam"] + (ymres - ypres - self.tsdiam)
 
-        self.lb_probe_d.set_text("%.4f" % diam)
+        self.display_result_d(diam)
         # move Z to start point up
         if self.z_clearance_up() == -1:
             return
@@ -2149,7 +1554,7 @@ class ProbeScreenClass(ProbeScreenBase):
         if self.ocode("o<psng_down> call") == -1:
             return
         a = self.probed_position_with_offsets()
-        self.lb_probe_z.set_text("%.4f" % float(a[2]))
+        self.display_result_z(float(a[2]))
         self.add_history(
             gtkbutton.get_tooltip_text(), "Z", 0, 0, 0, 0, 0, 0, 0, 0, a[2], 0, 0
         )
@@ -2207,8 +1612,6 @@ class ProbeScreenClass(ProbeScreenBase):
             self.prefs.getpref("use_toolmeasurement", False, bool)
         )
 
-        self.chk_auto_rott = self.builder.get_object("chk_auto_rott")
-        self.chk_auto_rott.set_active(self.prefs.getpref("chk_auto_rott", False, bool))
         self.xpym = self.builder.get_object("xpym")
         self.ym = self.builder.get_object("ym")
         self.xmym = self.builder.get_object("xmym")
@@ -2222,31 +1625,7 @@ class ProbeScreenClass(ProbeScreenBase):
         self.hole = self.builder.get_object("hole")
         self.angle = self.builder.get_object("angle")
 
-        self.spbtn1_search_vel = self.builder.get_object("spbtn1_search_vel")
-        self.spbtn1_probe_vel = self.builder.get_object("spbtn1_probe_vel")
-        self.spbtn1_z_clearance = self.builder.get_object("spbtn1_z_clearance")
-        self.spbtn1_probe_max = self.builder.get_object("spbtn1_probe_max")
-        self.spbtn1_probe_latch = self.builder.get_object("spbtn1_probe_latch")
-        self.spbtn1_probe_diam = self.builder.get_object("spbtn1_probe_diam")
-        self.spbtn1_xy_clearance = self.builder.get_object("spbtn1_xy_clearance")
-        self.spbtn1_edge_lenght = self.builder.get_object("spbtn1_edge_lenght")
-
         self.hal_led_set_m6 = self.builder.get_object("hal_led_set_m6")
-        self.hal_led_auto_rott = self.builder.get_object("hal_led_auto_rott")
-
-        self.spbtn_offs_angle = self.builder.get_object("spbtn_offs_angle")
-
-        self.lb_probe_xp = self.builder.get_object("lb_probe_xp")
-        self.lb_probe_yp = self.builder.get_object("lb_probe_yp")
-        self.lb_probe_xm = self.builder.get_object("lb_probe_xm")
-        self.lb_probe_ym = self.builder.get_object("lb_probe_ym")
-        self.lb_probe_lx = self.builder.get_object("lb_probe_lx")
-        self.lb_probe_ly = self.builder.get_object("lb_probe_ly")
-        self.lb_probe_z = self.builder.get_object("lb_probe_z")
-        self.lb_probe_d = self.builder.get_object("lb_probe_d")
-        self.lb_probe_xc = self.builder.get_object("lb_probe_xc")
-        self.lb_probe_yc = self.builder.get_object("lb_probe_yc")
-        self.lb_probe_a = self.builder.get_object("lb_probe_a")
 
         self.lx_out = self.builder.get_object("lx_out")
         self.lx_in = self.builder.get_object("lx_in")
@@ -2254,67 +1633,13 @@ class ProbeScreenClass(ProbeScreenBase):
         self.ly_in = self.builder.get_object("ly_in")
         self.tool_dia = self.builder.get_object("tool_dia")
 
-        self.halcomp.newpin("ps_searchvel", hal.HAL_FLOAT, hal.HAL_OUT)
-        self.halcomp.newpin("ps_probevel", hal.HAL_FLOAT, hal.HAL_OUT)
-        self.halcomp.newpin("ps_z_clearance", hal.HAL_FLOAT, hal.HAL_OUT)
-        self.halcomp.newpin("ps_probe_max", hal.HAL_FLOAT, hal.HAL_OUT)
-        self.halcomp.newpin("ps_probe_latch", hal.HAL_FLOAT, hal.HAL_OUT)
-        self.halcomp.newpin("ps_probe_diam", hal.HAL_FLOAT, hal.HAL_OUT)
-        self.halcomp.newpin("ps_xy_clearance", hal.HAL_FLOAT, hal.HAL_OUT)
-        self.halcomp.newpin("ps_edge_lenght", hal.HAL_FLOAT, hal.HAL_OUT)
-        self.halcomp.newpin("ps_offs_angle", hal.HAL_FLOAT, hal.HAL_OUT)
         self.halcomp.newpin("use_toolmeasurement", hal.HAL_BIT, hal.HAL_OUT)
         if self.chk_use_tool_measurement.get_active():
             self.halcomp["use_toolmeasurement"] = True
             self.hal_led_set_m6.hal_pin.set(1)
-        self.halcomp.newpin("auto_rott", hal.HAL_BIT, hal.HAL_OUT)
-        if self.chk_auto_rott.get_active():
-            self.halcomp["auto_rott"] = True
-            self.hal_led_auto_rott.hal_pin.set(1)
+
         self.halcomp.newpin("ps_error", hal.HAL_FLOAT, hal.HAL_OUT)
 
-        if self.inifile.find("TRAJ", "LINEAR_UNITS") not in ["metric", "mm"]:
-            # default values for inches
-            tup = (20.0, 2.0, 0.5, 1.0, 0.1, 0.125, 1.0, 1.25)
-        else:
-            tup = (300.0, 10.0, 3.0, 1.0, 0.5, 2.0, 5.0, 5.0)
-
-        self.spbtn1_search_vel.set_value(
-            self.prefs.getpref("ps_searchvel", tup[0], float)
-        )
-        self.spbtn1_probe_vel.set_value(
-            self.prefs.getpref("ps_probevel", tup[1], float)
-        )
-        self.spbtn1_z_clearance.set_value(
-            self.prefs.getpref("ps_z_clearance", tup[2], float)
-        )
-        self.spbtn1_probe_max.set_value(
-            self.prefs.getpref("ps_probe_max", tup[3], float)
-        )
-        self.spbtn1_probe_latch.set_value(
-            self.prefs.getpref("ps_probe_latch", tup[4], float)
-        )
-        self.spbtn1_probe_diam.set_value(
-            self.prefs.getpref("ps_probe_diam", tup[5], float)
-        )
-        self.spbtn1_xy_clearance.set_value(
-            self.prefs.getpref("ps_xy_clearance", tup[6], float)
-        )
-        self.spbtn1_edge_lenght.set_value(
-            self.prefs.getpref("ps_edge_lenght", tup[7], float)
-        )
-
-        self.spbtn_offs_angle.set_value(self.prefs.getpref("ps_offs_angle", 0.0, float))
-
-        self.halcomp["ps_searchvel"] = self.spbtn1_search_vel.get_value()
-        self.halcomp["ps_probevel"] = self.spbtn1_probe_vel.get_value()
-        self.halcomp["ps_z_clearance"] = self.spbtn1_z_clearance.get_value()
-        self.halcomp["ps_probe_max"] = self.spbtn1_probe_max.get_value()
-        self.halcomp["ps_probe_latch"] = self.spbtn1_probe_latch.get_value()
-        self.halcomp["ps_probe_diam"] = self.spbtn1_probe_diam.get_value()
-        self.halcomp["ps_xy_clearance"] = self.spbtn1_xy_clearance.get_value()
-        self.halcomp["ps_edge_lenght"] = self.spbtn1_edge_lenght.get_value()
-        self.halcomp["ps_offs_angle"] = self.spbtn_offs_angle.get_value()
         self.halcomp["ps_error"] = 0.0
 
         # For Auto Tool Measurement
